refactor(settings): collapse the settings-frontend slice modules into one outbound registry
The nine non-LaTeX `settingsView/frontend/slices/*` modules each existed only
to export a `satisfies Partial<SettingsViewOutboundHandlerRegistry>` fragment
that `messageDispatcher.ts` immediately spread back together. Every slice had
exactly one consumer, so the split bought a nine-file hop between "which
command is this" and "what does it do".

Move the handler bodies verbatim into the single `settingsViewHandlers`
literal in `messageDispatcher.ts`, grouped by domain comments, and delete the
nine modules. The registry keeps its name, module and
`SettingsViewOutboundHandlerRegistry` annotation, so exhaustiveness is still a
compile error and `SettingsApp.ts` / `SubscriptionUsageSettings.vitest.ts`
are untouched.

`slices/latexSlice.ts` stays a module (and a spread) for now: its arm is being
rewritten by the in-flight LaTeX wire-projection change, so it folds in once
that lands.

Behaviour-preserving: no wire, schema, or signal change. Net -110 lines,
-9 modules, -12 exported symbols.
